Move tests/test_base.py into tests/test_mapreduce.py
